Use a new base address for boot.img
The address 0x12000000 is being used as fastboot buffer in U-Boot, so we
need to use a new base address.

The mkbootimg tool uses 0x10000000 as the default which matches i.MX6Q
DRAM base address.  So in theory, we do not need to specify an address
but just use the default one.  However, U-Boot is using the address to
decide whether it executes the image in place.  With the default one,
U-Boot will launch the kernel with itself, ramdisk and fdt, in place.
It will cause a problem that ramdisk will be partially overwritten by
zImage decompressing, because ramdisk sits too close to zImage in
boot.img.  The problem is always there no matter which address we load
boot.img to, unless we get the image not XIP.  U-Boot will only relocate
ramdisk and fdt to high address when environment initrd_high and
fdt_high are defined.  That said, the load address of ramdisk and fdt in
boot.img are not respected by U-Boot, and only the load address of
kernel is.  We choose to specify base address 0x10800000 here, so that
when boot.img gets loaded to 0x12000000 by U-Boot, ramdisk and fdt will
be in place while kernel will be relocated to 0x10808000.  In such case,
ramdisk is far away from kernel image and will not be overwritten by
zImage decompressing.

Signed-off-by: Shawn Guo <shawn.guo@linaro.org>